Add handling for PENDING_CHARGE state
Based on testing the state is reached when
the hardware's charge limit blocks further
charging.

GNOME displays the state as "Not Charging"
in the power options and the quick settings.
The used icon is the default, non-charging
battery.

Currently batime shows PENDING_CHARGE as
"Estimating..." which suggests a temporary
state. Without external change
the state is permanent (based on my testing).